+++ app/src/main/res/layout/activity_main.xml
@@ -17,13 +17,10 @@
         app:layout_constraintTop_toTopOf="parent"
         app:title="@string/toolbar_title" />
 
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerView"
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/fragment_container"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/toolbar" />
+        android:layout_height="match_parent" />
 
 
 </LinearLayout>
